% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{arrange_taxa_pq}
\alias{arrange_taxa_pq}
\title{Arrange taxa by column values}
\usage{
arrange_taxa_pq(physeq, ..., clean_phyloseq_object = TRUE)
}
\arguments{
\item{physeq}{(phyloseq, required) A phyloseq object.}

\item{...}{\if{html}{\out{<data-masking>}} Variables to sort by. Use \code{desc()} for descending
order. Use \code{.} to refer to the phyloseq object.}

\item{clean_phyloseq_object}{if TRUE (default), the resulting phyloseq object
is cleaned using \code{clean_pq()} to remove empty taxa/samples.}
}
\value{
A phyloseq object with reordered taxa.
}
\description{
Reorder taxa based on tax_table columns or computed values. Supports the \code{.}
pronoun to refer to the phyloseq object for sorting by abundance.
}
\examples{
library(MiscMetabar)
# Arrange by taxonomy
arrange_taxa_pq(data_fungi, Phylum, Class)@tax_table[, "Phylum"]

# Arrange by total abundance (descending)
arrange_taxa_pq(data_fungi, dplyr::desc(taxa_sums(.)))

# order of columns matters
dfm_arr <- arrange_taxa_pq(data_fungi_mini, Class, Genus)@tax_table[, c("Class", "Genus")]
arrange_taxa_pq(data_fungi_mini, Genus, Class)@tax_table[, c("Class", "Genus")]



}
\author{
Adrien Taudière
}
