% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contam_tbl.R
\name{autoplot.contam_tbl}
\alias{autoplot.contam_tbl}
\alias{plot.contam_tbl}
\title{Diagnostic plot for a \code{contam_tbl}}
\usage{
\method{autoplot}{contam_tbl}(object, ...)

\method{plot}{contam_tbl}(x, ...)
}
\arguments{
\item{object, x}{(contam_tbl, required) A detector output.}

\item{...}{Currently ignored.}
}
\value{
A \code{ggplot} object (\code{autoplot}) or \code{x} invisibly (\code{plot}).
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


A lightweight overview of which detectors flagged how many taxa. For
publication-grade contamination figures, richer plotting lives in the
\code{ggplotpq} package and consumes the same \code{contam_tbl}.
}
\author{
Adrien Taudière
}
