% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contam_blocklist.R
\name{known_contaminant_genera}
\alias{known_contaminant_genera}
\title{Known reagent- and laboratory-contaminant genera}
\usage{
known_contaminant_genera()
}
\value{
A character vector of genus names.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


A curated vector of bacterial genera repeatedly reported as reagent,
kit and laboratory contaminants in low-biomass sequencing studies. The
list is compiled from Salter et al. (2014), the \code{micRoclean} workflow
and the GRIMER \code{contaminants.yml} / \code{human-related.yml} resources. It is
intended as a starting blocklist for \code{\link[=contam_blocklist_pq]{contam_blocklist_pq()}}; users
should adapt it to their own substrate and primers.
}
\examples{
head(known_contaminant_genera())
length(known_contaminant_genera())
}
\references{
Salter SJ, Cox MJ, Turek EM, et al. (2014) Reagent and laboratory
contamination can critically impact sequence-based microbiome analyses.
BMC Biology 12:87. \doi{10.1186/s12915-014-0087-z}
}
\author{
Adrien Taudière
}
