% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{mutate_samdata_pq}
\alias{mutate_samdata_pq}
\title{Add or modify columns in sample_data}
\usage{
mutate_samdata_pq(physeq, ...)
}
\arguments{
\item{physeq}{(phyloseq, required) A phyloseq object.}

\item{...}{\if{html}{\out{<data-masking>}} Name-value pairs. The name gives the name of the
column in the output. The value must be a vector of length 1 (recycled) or
exactly the same length as the number of samples. Use \code{.} to refer to the

phyloseq object.}
}
\value{
A phyloseq object with modified sample_data (same samples, modified
or new columns).
}
\description{
Create new columns or modify existing ones in sample_data using data masking.
Supports the \code{.} pronoun to refer to the phyloseq object.

This function only modifies the sample_data slot (columns/metadata). It cannot
add or remove samples. The number of samples and sample names are preserved.

Unlike \code{dplyr::mutate()}, columns created in the same call cannot reference
each other (e.g., \code{mutate_samdata_pq(pq, a = 1, b = a + 1)} will not work
because \code{a} is not yet available when \code{b} is evaluated).
}
\examples{
library(MiscMetabar)
# Add a new column based on sequencing depth
mutate_samdata_pq(data_fungi, log_depth = log(sample_sums(.)))

# Modify an existing column
mutate_samdata_pq(data_fungi, Height = toupper(Height))
}
\author{
Adrien Taudière
}
