% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contam_tbl.R
\name{print.contam_tbl}
\alias{print.contam_tbl}
\title{Print a \code{contam_tbl}}
\usage{
\method{print}{contam_tbl}(x, ...)
}
\arguments{
\item{x}{(contam_tbl, required) The object to print.}

\item{...}{Passed to the tibble print method.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{contam_tbl}
}
\author{
Adrien Taudière
}
