% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{rename_samples_pq}
\alias{rename_samples_pq}
\title{Rename columns in sample_data}
\usage{
rename_samples_pq(physeq, ...)
}
\arguments{
\item{physeq}{(phyloseq, required) A phyloseq object.}

\item{...}{\if{html}{\out{<tidy-select>}} Name-value pairs where the name is the new name
and the value is the old name. Use \code{new_name = old_name} syntax.}
}
\value{
A phyloseq object with renamed sample_data columns.
}
\description{
Rename columns in sample_data using tidyselect semantics.
}
\examples{
library(MiscMetabar)
# Rename a single column
rename_samples_pq(data_fungi, sample_height = Height)

# Rename multiple columns
rename_samples_pq(data_fungi, sample_height = Height, sample_time = Time)
}
\author{
Adrien Taudière
}
