% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{select_samdata_pq}
\alias{select_samdata_pq}
\title{Select columns from sample_data in a phyloseq object}
\usage{
select_samdata_pq(physeq, ...)
}
\arguments{
\item{physeq}{(phyloseq, required) A phyloseq object.}

\item{...}{\if{html}{\out{<tidy-select>}} One or more unquoted expressions separated by
commas. Variable names can be used as if they were positions in the data
frame, so expressions like \code{x:y} can be used to select a range of
variables.}
}
\value{
A phyloseq object with selected sample_data columns.
}
\description{
Select sample_data columns using tidyselect semantics.
}
\examples{
library(MiscMetabar)
# Select specific columns
select_samdata_pq(data_fungi, Height, Time)

# Select a range of columns
select_samdata_pq(data_fungi, Height:Time)

# Exclude columns
select_samdata_pq(data_fungi, !Sample_id)
}
\author{
Adrien Taudière
}
