% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{select_taxa_pq}
\alias{select_taxa_pq}
\title{Select columns from tax_table in a phyloseq object}
\usage{
select_taxa_pq(physeq, ...)
}
\arguments{
\item{physeq}{(phyloseq, required) A phyloseq object.}

\item{...}{\if{html}{\out{<tidy-select>}} One or more unquoted expressions separated by
commas. Variable names can be used as if they were positions in the data
frame, so expressions like \code{Kingdom:Genus} can be used to select a range
of taxonomic ranks.}
}
\value{
A phyloseq object with selected tax_table columns.
}
\description{
Select tax_table columns (taxonomic ranks) using tidyselect semantics.
}
\examples{
library(MiscMetabar)
# Select specific ranks
select_taxa_pq(data_fungi, Kingdom, Phylum, Class)

# Select a range of ranks
select_taxa_pq(data_fungi, Kingdom:Genus)

# Exclude ranks
select_taxa_pq(data_fungi, !Species)
}
\author{
Adrien Taudière
}
