% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{slice_samples_pq}
\alias{slice_samples_pq}
\title{Subset samples by position}
\usage{
slice_samples_pq(physeq, ..., clean_phyloseq_object = TRUE)
}
\arguments{
\item{physeq}{(phyloseq, required) A phyloseq object.}

\item{...}{\if{html}{\out{<data-masking>}} Integer row indices. Positive values select
samples, negative values drop samples.}

\item{clean_phyloseq_object}{if TRUE (default), the resulting phyloseq object
is cleaned using \code{clean_pq()} to remove empty taxa/samples.}
}
\value{
A phyloseq object with selected samples.
}
\description{
Select samples by their integer positions.
}
\examples{
library(MiscMetabar)
# Select first 5 samples
slice_samples_pq(data_fungi, 1:5)

# Remove first 2 samples
slice_samples_pq(data_fungi, -(1:2))
}
\author{
Adrien Taudière
}
