% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{slice_taxa_pq}
\alias{slice_taxa_pq}
\title{Subset taxa by position}
\usage{
slice_taxa_pq(physeq, ..., clean_phyloseq_object = TRUE)
}
\arguments{
\item{physeq}{(phyloseq, required) A phyloseq object.}

\item{...}{\if{html}{\out{<data-masking>}} Integer row indices. Positive values select
taxa, negative values drop taxa.}

\item{clean_phyloseq_object}{if TRUE (default), the resulting phyloseq object
is cleaned using \code{clean_pq()} to remove empty taxa/samples.}
}
\value{
A phyloseq object with selected taxa.
}
\description{
Select taxa by their integer positions.
}
\examples{
library(MiscMetabar)
# Select first 10 taxa
slice_taxa_pq(data_fungi, 1:10)

# Remove first 5 taxa
slice_taxa_pq(data_fungi, -(1:5))
}
\author{
Adrien Taudière
}
